+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 - New #83: Add `answerShippingQuery` and `answerPreCheckoutQuery` methods, and `ShippingOption` type.
 - New #84: Add `refundStarPayment` method.
 - New #85: Add `setPassportDataErrors` method and `PassportElementError*` types.
+- New #86: Add `sendGame` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.